Switching to caching based on reflection
